Move F680 application submit API tests to pytest-style
