+++ api/f680/exporter/tests/test_views.py
@@ -16,7 +16,12 @@ from api.f680.tests.factories import (
     SubmittedF680ApplicationFactory,  # /PS-IGNORE
 )
 
+pytest_plugins = [
+    "api.tests.unit.fixtures.core",
+]
 
+
+# TODO: Move the below DataTestClient tests over to pytest style (class below)
 class F680ApplicationViewSetTests(DataTestClient):  # /PS-IGNORE
     def setUp(self):
         super().setUp()
@@ -266,57 +271,62 @@ class F680ApplicationViewSetTests(DataTestClient):  # /PS-IGNORE
         }
         self.assertEqual(response.data, expected_result)
 
-    def test_POST_submit_empty_data_not_found(self):
-        url = reverse("exporter_f680:application_submit", kwargs={"f680_application_id": uuid4()})
-        response = self.client.post(url, **self.exporter_headers)
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
-    def test_POST_submit_submitted_application_not_found(self):
-        f680_application = SubmittedF680ApplicationFactory(organisation=self.organisation)
-        url = reverse("exporter_f680:application_submit", kwargs={"f680_application_id": f680_application.id})
-        response = self.client.post(url, **self.exporter_headers)
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+class TestF680ApplicationViewSet:
 
     @freeze_time("2025-01-01 12:00:01")
-    def test_POST_submit_success(self):
+    def test_POST_submit_success(
+        self, api_client, exporter_headers, exporter_user, organisation, data_application_json
+    ):
         f680_application = F680ApplicationFactory(
-            organisation=self.organisation,
-            application={"old key": "old value"},
+            organisation=organisation,
+            application=data_application_json,
             reference_code=None,
         )
         url = reverse("exporter_f680:application_submit", kwargs={"f680_application_id": f680_application.id})
-        response = self.client.post(url, **self.exporter_headers)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        response = api_client.post(url, **exporter_headers)
+        assert response.status_code == status.HTTP_200_OK
         f680_application.refresh_from_db()
         assert f680_application.status.status == "submitted"
         assert f680_application.submitted_at == timezone.now()
         assert f680_application.sla_days == 0
-        assert f680_application.submitted_by == self.exporter_user
+        assert f680_application.submitted_by == exporter_user
         assert f680_application.reference_code.startswith("F680")
 
         expected_result = {
             "id": str(f680_application.id),
-            "application": {"old key": "old value"},
+            "application": data_application_json,
             "reference_code": f680_application.reference_code,
             "organisation": {
-                "id": str(self.organisation.id),
-                "name": self.organisation.name,
-                "type": self.organisation.type,
-                "status": self.organisation.status,
+                "id": str(organisation.id),
+                "name": organisation.name,
+                "type": organisation.type,
+                "status": organisation.status,
             },
             "submitted_at": timezone.now().strftime("%Y-%m-%dT%H:%M:%SZ"),
             "submitted_by": {
-                "email": self.exporter_user.email,
-                "first_name": self.exporter_user.first_name,
-                "id": str(self.exporter_user.baseuser_ptr_id),
-                "last_name": self.exporter_user.last_name,
-                "pending": self.exporter_user.pending,
+                "email": exporter_user.email,
+                "first_name": exporter_user.first_name,
+                "id": str(exporter_user.baseuser_ptr_id),
+                "last_name": exporter_user.last_name,
+                "pending": exporter_user.pending,
             },
         }
-        self.assertEqual(response.data, expected_result)
+        assert response.data == expected_result
 
-    def test_POST_submit_different_organisation_not_found(self):
+    def test_POST_submit_different_organisation_not_found(self, api_client, exporter_headers):
         f680_application = SubmittedF680ApplicationFactory()
         url = reverse("exporter_f680:application_submit", kwargs={"f680_application_id": f680_application.id})
-        response = self.client.post(url, {"application": {"new": "new value"}}, **self.exporter_headers)
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+        response = api_client.post(url, {"application": {"new": "new value"}}, **exporter_headers)
+        assert response.status_code == status.HTTP_404_NOT_FOUND
+
+    def test_POST_submit_empty_data_not_found(self, api_client, exporter_headers):
+        url = reverse("exporter_f680:application_submit", kwargs={"f680_application_id": uuid4()})
+        response = api_client.post(url, **exporter_headers)
+        assert response.status_code == status.HTTP_404_NOT_FOUND
+
+    def test_POST_submit_submitted_application_not_found(self, api_client, exporter_headers, organisation):
+        f680_application = SubmittedF680ApplicationFactory(organisation=organisation)
+        url = reverse("exporter_f680:application_submit", kwargs={"f680_application_id": f680_application.id})
+        response = api_client.post(url, **exporter_headers)
+        assert response.status_code == status.HTTP_404_NOT_FOUND
